Cleanup, config and Menus
Changelog:
- Fully implemented config, through the game
- Added labels to buttons in menu
- Added 2 more icons to menu
- Massive code cleanup
- Added logic for adding new Shops.
- Fixed logic for creating buttons for new shops.
- Minor fixes here and there

This update added various bug fixes, whilst making an effort to clean up the source code, for future development. It also added the logic for creating new shops, and having buttons that work. It has made it possible to begin on Shop rework, and then the first Beta is ready.